# Conflicts:
#	docs/config.toml
#	docs/content.zh/docs/connectors/table/filesystem.md
#	docs/content.zh/docs/dev/python/debugging.md
#	docs/content.zh/docs/dev/python/dependency_management.md
#	docs/content.zh/docs/dev/python/python_config.md
#	docs/content.zh/docs/dev/python/table/intro_to_table_api.md
#	docs/content.zh/docs/dev/python/table/operations/row_based_operations.md
#	docs/content.zh/docs/dev/python/table/udfs/python_udfs.md
#	docs/content.zh/docs/dev/table/common.md
#	docs/content.zh/docs/dev/table/concepts/timezone.md
#	docs/content.zh/docs/dev/table/data_stream_api.md
#	docs/content.zh/docs/dev/table/sql/queries/joins.md
#	docs/content.zh/docs/dev/table/sql/queries/over-agg.md
#	docs/content.zh/docs/dev/table/sql/queries/overview.md
#	docs/content.zh/docs/dev/table/sql/set.md
#	docs/content.zh/docs/dev/table/sqlClient.md
#	docs/content.zh/docs/dev/table/tuning.md
#	docs/content.zh/docs/flinkDev/ide_setup.md
#	docs/content.zh/docs/libs/gelly/bipartite_graph.md
#	docs/content.zh/docs/libs/gelly/graph_api.md
#	docs/content/docs/dev/python/debugging.md
#	docs/content/docs/dev/python/dependency_management.md
#	docs/content/docs/dev/python/python_config.md
#	docs/content/docs/dev/python/table/intro_to_table_api.md
#	docs/content/docs/dev/python/table/operations/row_based_operations.md
#	docs/content/docs/dev/python/table/table_environment.md
#	docs/content/docs/dev/python/table/udfs/python_udfs.md
#	docs/content/docs/dev/table/concepts/timezone.md
#	docs/content/docs/dev/table/data_stream_api.md
#	docs/content/docs/dev/table/sql/queries/over-agg.md
#	docs/content/docs/dev/table/sql/queries/overview.md
#	docs/content/docs/dev/table/sql/set.md
#	docs/content/docs/dev/table/sqlClient.md
#	docs/content/docs/libs/gelly/bipartite_graph.md
#	docs/content/docs/libs/gelly/graph_api.md
#	docs/layouts/shortcodes/generated/table_config_configuration.html
#	flink-connectors/flink-connector-jdbc/pom.xml
#	flink-connectors/flink-connector-jdbc/src/test/java/org/apache/flink/connector/jdbc/xa/JdbcExactlyOnceSinkE2eTest.java
#	flink-kubernetes/src/main/java/org/apache/flink/kubernetes/highavailability/KubernetesHaServices.java
#	flink-kubernetes/src/test/java/org/apache/flink/kubernetes/highavailability/KubernetesHaServicesTest.java
#	flink-python/pyflink/datastream/tests/test_stream_execution_environment.py
#	flink-python/pyflink/fn_execution/coder_impl_fast.pxd
#	flink-python/pyflink/fn_execution/coder_impl_fast.pyx
#	flink-python/pyflink/fn_execution/flink_fn_execution_pb2.py
#	flink-python/pyflink/fn_execution/tests/test_fast_coders.py
#	flink-python/pyflink/table/tests/test_pandas_udf.py
#	flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/channel/RecoveredChannelStateHandler.java
#	flink-runtime/src/main/java/org/apache/flink/runtime/highavailability/AbstractHaServices.java
#	flink-runtime/src/main/java/org/apache/flink/runtime/highavailability/zookeeper/ZooKeeperHaServices.java
#	flink-runtime/src/test/java/org/apache/flink/runtime/dispatcher/FileExecutionGraphInfoStoreTest.java
#	flink-runtime/src/test/java/org/apache/flink/runtime/highavailability/AbstractHaServicesTest.java
#	flink-runtime/src/test/java/org/apache/flink/runtime/highavailability/zookeeper/ZooKeeperHaServicesTest.java
#	flink-runtime/src/test/java/org/apache/flink/runtime/scheduler/adaptive/AdaptiveSchedulerClusterITCase.java
#	flink-runtime/src/test/java/org/apache/flink/runtime/scheduler/adaptive/AdaptiveSchedulerSimpleITCase.java
#	flink-runtime/src/test/scala/org/apache/flink/runtime/jobmanager/Tasks.scala
#	flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/StreamTask.java
#	flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/SubtaskCheckpointCoordinatorImpl.java
#	flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/tasks/SourceOperatorStreamTaskTest.java
#	flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/tasks/StreamTaskTest.java
#	flink-table/flink-sql-client/src/test/java/org/apache/flink/table/client/SqlClientTest.java
#	flink-table/flink-sql-client/src/test/resources/sql/select.q
#	flink-table/flink-sql-parser/src/main/codegen/data/Parser.tdd
#	flink-table/flink-sql-parser/src/test/java/org/apache/flink/sql/parser/FlinkSqlParserImplTest.java
#	flink-table/flink-table-api-java-bridge/src/main/java/org/apache/flink/table/api/bridge/java/StreamTableEnvironment.java
#	flink-table/flink-table-api-java-bridge/src/test/java/org/apache/flink/table/factories/DataGenTableSourceFactoryTest.java
#	flink-table/flink-table-api-scala-bridge/src/main/scala/org/apache/flink/table/api/bridge/scala/StreamTableEnvironment.scala
#	flink-table/flink-table-common/src/main/java/org/apache/flink/table/factories/FactoryUtil.java
#	flink-table/flink-table-common/src/main/java/org/apache/flink/table/functions/BuiltInFunctionDefinitions.java
#	flink-table/flink-table-common/src/main/java/org/apache/flink/table/types/inference/TypeStrategies.java
#	flink-table/flink-table-common/src/test/java/org/apache/flink/table/factories/FactoryUtilTest.java
#	flink-table/flink-table-planner-blink/src/main/java/org/apache/flink/table/planner/operations/SqlToOperationConverter.java
#	flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/plan/utils/WindowJoinUtil.scala
#	flink-table/flink-table-planner-blink/src/test/java/org/apache/flink/table/planner/factories/TestValuesRuntimeFunctions.java
#	flink-table/flink-table-planner-blink/src/test/java/org/apache/flink/table/planner/operations/SqlToOperationConverterTest.java
#	flink-table/flink-table-planner-blink/src/test/resources/org/apache/flink/table/planner/plan/stream/sql/join/WindowJoinTest.xml
#	flink-table/flink-table-planner-blink/src/test/scala/org/apache/flink/table/planner/runtime/stream/sql/CalcITCase.scala
#	flink-table/flink-table-planner/src/main/java/org/apache/flink/table/parse/ExtendedParser.java
#	flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/ParserImpl.java
#	flink-table/flink-table-planner/src/main/scala/org/apache/flink/table/api/internal/TableEnvImpl.scala
#	flink-table/flink-table-planner/src/test/scala/org/apache/flink/table/util/python/PythonTableUtilsTest.scala
#	flink-table/flink-table-runtime-blink/src/main/java/org/apache/flink/table/filesystem/stream/PartitionTimeCommitTrigger.java
#	flink-test-utils-parent/flink-test-utils-junit/src/main/java/org/apache/flink/testutils/junit/RetryRule.java
#	flink-tests/src/test/java/org/apache/flink/test/scheduling/ReactiveModeITCase.java
#	pom.xml
#	tools/azure-pipelines/jobs-template.yml

+++ flink-kubernetes/src/test/java/org/apache/flink/kubernetes/KubernetesResourceManagerDriverTest.java
@@ -362,7 +362,12 @@ public class KubernetesResourceManagerDriverTest
                                             .getMebiBytes())));
             assertThat(
                     resourceRequirements.getRequests().get(Constants.RESOURCE_NAME_CPU).getAmount(),
-                    is(String.valueOf(taskExecutorProcessSpec.getCpuCores().getValue())));
+                    is(
+                            String.valueOf(
+                                    taskExecutorProcessSpec
+                                            .getCpuCores()
+                                            .getValue()
+                                            .doubleValue())));
 
             assertThat(
                     resourceRequirements
@@ -376,7 +381,12 @@ public class KubernetesResourceManagerDriverTest
                                             .getMebiBytes())));
             assertThat(
                     resourceRequirements.getLimits().get(Constants.RESOURCE_NAME_CPU).getAmount(),
-                    is(String.valueOf(taskExecutorProcessSpec.getCpuCores().getValue())));
+                    is(
+                            String.valueOf(
+                                    taskExecutorProcessSpec
+                                            .getCpuCores()
+                                            .getValue()
+                                            .doubleValue())));
         }
 
         @Override
